backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.7.2
astroid==3.3.5
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
deprecation==2.1.0
dill==0.3.9
distlib==0.3.9
distro==1.9.0
dj-database-url==2.1.0
Django==5.0.1
django-cors-headers==4.3.1
django-filter==24.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
environs==11.0.0
filelock==3.16.1
gotrue==2.9.1
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.25.2
hyperframe==6.0.1
idna==3.10
inflection==0.5.1
isort==5.13.2
jedi==0.19.1
jiter==0.6.1
marshmallow==3.21.1
mccabe==0.7.0
mysqlclient==2.2.4
openai==1.51.2
packaging==24.0
parso==0.8.4
pillow==10.2.0
platformdirs==4.3.6
postgrest==0.15.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pylint==3.3.1
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6_sip==13.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
qrcode==8.0
realtime==1.0.6
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.4.4
storage3==0.7.7
StrEnum==0.4.15
supabase==2.3.4
supafunc==0.3.3
tomlkit==0.13.2
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.26.6
websockets==12.0
